seperated file atomic to atomicGeneral and groupTerm
