+++ pkg/symbolicexpr/groupTerm.go
@@ -2,7 +2,6 @@ package symbolicexpr
 
 import (
 	"fmt"
-	"slices"
 
 	"github.com/np-guard/vmware-analyzer/pkg/collector"
 )
@@ -74,55 +73,4 @@ func (groupTerm groupAtomicTerm) disjoint(otherAt atomic, hints *Hints) bool {
 // todo: perhaps this can have a general implementation instead of struct specific one
 func (groupTerm groupAtomicTerm) supersetOf(otherAt atomic, hints *Hints) bool {
 	return hints.disjoint(groupTerm.name(), otherAt.name()) && groupTerm.isNegation() && !otherAt.isNegation()
-}
-
-func (tautology) string() string {
-	return "*"
-}
-
-func (tautology) name() string {
-	return ""
-}
-
-func (tautology) negate() atomic {
-	return tautology{}
-}
-
-func (tautology) isNegation() bool {
-	return false
-}
-
-func (tautology) IsTautology() bool {
-	return true
-}
-
-// returns true iff otherAt is negation of
-// once we cache the atomic terms, we can just compare pointers
-func (tautology) isNegateOf(atomic) bool {
-	return false
-}
-func (tautology) AsSelector() (string, bool) {
-	return "", false
-}
-
-// tautology is not disjoint to any atomic term
-func (tautology) disjoint(atomic, *Hints) bool {
-	return false
-}
-
-func (tautology) supersetOf(atom atomic, hints *Hints) bool {
-	return atom.IsTautology()
-}
-
-// are two given by name atomicTerms in disjoint list
-func (hints *Hints) disjoint(name1, name2 string) bool {
-	if name1 == name2 {
-		return false
-	}
-	for _, disjointGroup := range hints.GroupsDisjoint {
-		if slices.Contains(disjointGroup, name1) && slices.Contains(disjointGroup, name2) {
-			return true
-		}
-	}
-	return false
 }
